refactor: OverlaidPopup 위치 변경
